requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118

torch
numpy==1.26.4
sounddevice
playsound3
TTS
sentence-transformers
faiss-cpu
requests
whisperx
ollama
